Refactor Comps
