--- Source/Comps/CompProperties_Reincarnate.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using Verse;
-
-namespace Typhon
-{
-    internal class CompProperties_Reincarnate : CompProperties
-    {
-        public PawnKindDef pawnKind = null;
-        public int amount = 0;
-        public CompProperties_Reincarnate()
-        {
-            compClass = typeof(CompReincarnate);
-        }
-    }
-}
